Modified to use my hub.docker.com repo area (pmichali/news-service) and to use nodeport, as running on bare-metal and don't have a load balancer.
Also, Alpine failed to run app, as it was dynamically linking to
libs that are not at same place, and failed to run app. Solution was
to turn off dynamic linking when build app.

If run DinD for cluster, may need to stop dashboard deployment, as
the ports conflict (or change the port used here).

Currently cannot access this web server from outside of cluster.